bucket/requirements.txt
# Bucket Backend Requirements

# Core dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Firebase Admin SDK
firebase-admin>=6.0.0

# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1

# Embeddings
sentence-transformers>=2.2.0

# Vector store
chromadb>=0.4.0

# Document loaders
pdfplumber>=0.10.0
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.0
beautifulsoup4>=4.12.0

# Retrieval
rank-bm25>=0.2.2
